feature: This commit got abit out of hand, alot of refactoring made. New widgets introduced. Design changes.

+++ lib/pages/widgets/wind-direction.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'dart:math' as math;
 
 class WindDirection extends StatelessWidget {
-  final int angle;
+  final double angle;
   final double windSpeed;
   final double windGust;
